Webapp/requirements.txt
Werkzeug==0.15.5
Flask==1.1.1
numpy>=1.9.2
Keras==2.2.4
gevent
pillow
h5py
tensorflow==1.8.0
gunicorn==19.9.0
opencv-python==3.3.0.9
scikit-image
